refactor validators namespace
